requirements.txt
certifi==2023.5.7
babel==2.12.1
flask-babel==3.1.0
flask==2.3.2
jinja2==3.1.2
lxml==4.9.3
pygments==2.15.1
python-dateutil==2.8.2
pyyaml==6.0
httpx[http2]==0.21.2
Brotli==1.0.9
uvloop==0.17.0
httpx-socks[asyncio]==0.7.2
setproctitle==1.3.2
redis==4.6.0
markdown-it-py==3.0.0
typing_extensions==4.7.1
fasttext-predict==0.9.2.1
pytomlpp==1.0.13
python-dotenv==1.0.0
Flask-PyMongo==2.3.0